% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R, R/bbm_method.R
\docType{methods}
\name{bbm}
\alias{bbm}
\alias{bbm-methods}
\alias{bbm-method}
\alias{bbm,FLStock,ANY,ANY-method}
\alias{bbm,FLQuant,FLIndices,FLIndices-method}
\alias{bbm,FLQuant,FLQuant,FLQuant-method}
\title{Method bbm}
\usage{
bbm(object, indicesB, indicesP, ...)

\S4method{bbm}{FLQuant,FLQuants,FLQuants}(object, indicesB, indicesP,
  findicesB, findicesP, control, inits)

\S4method{bbm}{FLStock,ANY,ANY}(object, indicesB, indicesP,
  findicesB = NULL, findicesP = NULL, control, inits)

\S4method{bbm}{FLQuant,FLIndices,FLIndices}(object, indicesB, indicesP,
  findicesB = NULL, findicesP = NULL, control, inits)

\S4method{bbm}{FLQuant,FLQuant,FLQuant}(object, indicesB, indicesP,
  findicesB, findicesP, control, inits)
}
\arguments{
\item{object}{An \code{FLQuant} with catch information (in mass) or an \code{FLStock}. These objects must only have two age classes (recruits and adults)
and the number of seasons should be 1 or the number of seasons determined by the timing of the different indices.}

\item{indicesB}{Abundance indices in total biomass (element of class: \code{FLQuants}, \code{FLQuant} or \code{FLIndices}) from surveys. Please assign a survey name to each index.}

\item{indicesP}{Percentage of recruits in biomass (element of class: \code{FLQuants}, \code{FLQuant} or \code{FLIndices}) from surveys. Please assign a survey name to each proportion.}

\item{findicesB}{A \code{vector} with fraction of the year corresponding to each of the indicesB.}

\item{findicesP}{A \code{vector} with fraction of the year corresponding to each of the indicesP.}

\item{control}{A \code{bbmControl} with control arguments.}

\item{inits}{An \code{FLPar} with initial values.}
}
\value{
An object of class \code{bbmFit}.
}
\description{
Function to fit a two-stage biomass-based model.
}
\examples{

# Load required libraries
library(bbm)

# Load data
data(ane)

# Case:  object='FLQuant'; indicesB=indicesP='FLQuants'; control='bbmControl'; inits='FLPar'

run <- bbm( catch.ane, 
            indicesB=lapply( indicesB.ane, function(x) x@index), 
            indicesP=lapply( indicesP.ane, function(x) x@index), 
            findicesB=unlist(lapply( indicesB.ane, function(x) mean(range(x)[c('startf','endf')]))),
            findicesP=unlist(lapply( indicesP.ane, function(x) mean(range(x)[c('startf','endf')]))),
            control=control.ane, inits=inits.ane)
            
is(run)
slotNames(run)

# Run the assessment with fixed parameters
ctrl <- control.ane
ctrl@param.fix['q_depm'] <- 1

run0 <- bbm(catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=ctrl, inits=inits.ane)
params(run0)['q_depm']; inits.ane['q_depm']
params.se(run0)['q_depm']


# Case:  object='FLStock'; indicesB=indicesP='FLIndices'; control='bbmControl'; inits='FLPar'

stock <- FLStock(catch.n=catch.ane, catch.wt=catch.ane*0+1)
units(stock@catch.wt) <- ''
stock@catch <- quantSums(stock@catch.n*stock@catch.wt)

run2 <- bbm(stock, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)


# Case:  object='FLQuant'; indicesB=indicesP='FLIndices'; control='bbmControl'; inits='FLPar'

run3 <- bbm(catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane) 


# Case:  object='FLQuant'; indicesB=indicesP='FLQuant'; control='bbmControl'; inits='FLPar'

namdel <- c("q_acoustic","psi_acoustic","xi_acoustic") # we will take only one of the indices --> need to delete the parameters related to other indices
control <- control.ane
control@param.fix <- control@param.fix[dimnames(control@param.fix)$params[!dimnames(control@param.fix)$params \%in\% namdel],]
inits   <- inits.ane[dimnames(inits.ane)$params[!dimnames(inits.ane)$params \%in\% namdel],]
run4 <- bbm( catch.ane, indicesB=indicesB.ane[[1]]@index, indicesP=indicesP.ane[[1]]@index, 
             findicesB=c( depm=(indicesB.ane[[1]]@range[['startf']]+indicesB.ane[[1]]@range[['endf']])/2), 
             findicesP=c( depm=(indicesP.ane[[1]]@range[['startf']]+indicesP.ane[[1]]@range[['endf']])/2), 
             control=control, inits=inits)

# Plot assessed populations
biomass <- FLQuants()
runs <- c('run','run0','run2','run3','run4')
names(runs) <- c('bc','fixed_qdepm','run2','run3','only_depm')
for (i in 1:length(runs)) biomass[[i]] <- quantSums(stock.bio(get(runs[i])))
names(biomass) <- names(runs)
plot( biomass)

}
\seealso{
\link{bbmFit}, \linkS4class{FLQuant}, \linkS4class{FLQuants}, \linkS4class{bbmControl}, \linkS4class{FLPar}, \link{bbmFLPar}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{bbm}
\keyword{methods}
