% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R, R/bbmControl_class.R
\docType{class}
\name{bbmControl}
\alias{bbmControl}
\alias{bbmControl-methods}
\alias{bbmControl-class}
\alias{g,bbmControl-method}
\alias{param.fix,bbmControl-method}
\alias{bbmControl,missing-method}
\title{bbmControl class}
\usage{
bbmControl(object, ...)

\S4method{bbmControl}{missing}(object, ...)
}
\description{
\code{bbmControl} class is used for setting the control parameters for the \code{bbm} function.
}
\section{Slots}{

\describe{
\item{g        }{ Instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult),
\code{c(rec=numeric(1), adult=numeric(1))}.}
\item{param.fix}{ An \code{FLPar} with value 1 if the parameter is fixed at its initial value and 0 otherwise.}
}
}

\section{Validity}{

\describe{
\item{Dimensions}{\code{g} must be a vector of length 2 and with the names 'rec' and 'adult'}
\item{Class     }{\code{param.fix} must be of class \code{FLPar}}
}
You can inspect the class validity function by using \code{getValidity(getClassDef('bbmControl'))}
}

\section{Accessors}{

All slots in the class have accessor methods defined that allow retrieving individual slots.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity.
}

\examples{

# Load data
data(ane)

# Generate an object of FLPar class (different alternatives)
bbmControl()                     # empty object
slotNames(bbmControl())          # slots

bbmControl( g=c(rec=0.68, adult=0.68), param.fix=FLPar(nyear=20, nindex=3)) # setting values for the slots

# Run assessment (control must be of class bbmControl)
class(control.ane)
run <- bbm(catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)
run

}
\seealso{
\link{bbm}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez
}
\keyword{classes}
