% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBMfit.R
\docType{class}
\name{BBMfit}
\alias{BBMfit}
\alias{BBMfit-methods}
\alias{BBMfit-class}
\title{BBMfit class for storing the output of the BBM function.}
\description{
This includes abundance estimates in biomass (for recruits and adults) and information on the model fitting.
}
\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If \code{dim} or \code{dimnames} are provided, this is used
for sizing the stock.bio slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
\item{BBMfit   }{: Creation of a new BBMfit object.}
\item{stock    }{: Extracts information on estimated abundances in biomass for recruits and adults. Returns an \code{FLQuant}.}
\item{+        }{: Updates an \code{FLStock} with new information on the BBM assessment.}
\item{residuals}{: XXX.}
\item{xxx}{: XXX.}
\item{xxx}{: XXX.}
\item{xxx}{: XXX.}
}
}

\examples{

# Load data
data(ane)

# Run assessment (output is of class BBMfit)
run <- BBM(catch.ane, indices=indices.ane, control=control.ane, inits=inits.ane)
class(run)
run

# Available methods for the class
# - stock()
stock(run)
run@stock.bio

# - residuals
residuals(run)

# - AIC and BIC
AIC(run)
BIC(run)


}
\seealso{
\link{BBM}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez
}
\keyword{classes}
