% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R, R/bbmFitresiduals_class.R
\docType{class}
\name{bbmFitresiduals}
\alias{bbmFitresiduals}
\alias{bbmFitresiduals-methods}
\alias{bbmFitresiduals-class}
\alias{residualsB,bbmFitresiduals-method}
\alias{residualsP,bbmFitresiduals-method}
\alias{plot,bbmFitresiduals-method}
\alias{qqmath,bbmFitresiduals-method}
\title{bbmFitresiduals class}
\usage{
bbmFitresiduals(object, ...)
}
\description{
\code{bbmFitresiduals} class is used for storing the residuals for the fitted indices in the \code{bbm} function.
This includes residuals for indices in total biomass (indicesB) and as proportions of recruits in biomass (indicesP).
}
\section{Slots}{

\describe{
\item{residuals.B}{Residuals for indices in total biomass (indicesB), \code{FLQuants}.}
\item{residuals.P}{Residuals for indices as proportions of recruits in biomass (indicesP), \code{FLQuants}.}
}
}

\section{Accessors}{

All slots in the class have accessor methods defined that allow retrieving individual slots.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class. these are:

\describe{
\item{plot  }{Method to plot the Pearson residuals.}
\item{qqmath}{Method to do a qqplot of the Pearson residuals of survey indices.}
}
}

\examples{

# Load data
data(ane)

# New element of class 'bbmFitresiduals' (empty)
bbmFitresiduals()

obj <- bbm( catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)

res <- residuals(obj)
class(res)
slotNames(res)

# Starndardized residuals plots
plot(res)
qqmath(res)

}
\seealso{
\link{bbmFit}, \link{plot}, \link{qqmath}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez
}
\keyword{classes}
