% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmFLPar_fun.R
\name{bbmFLPar}
\alias{bbmFLPar}
\title{Function to generate an FLPar for bbm function}
\usage{
bbmFLPar(x = NULL, years, namesB, namesP, niter = 1,
  logscale = FALSE)
}
\arguments{
\item{x}{Input numeric values for the parameters. If not set, then NA value is assigned.}

\item{years}{Names of the years used for fitting.}

\item{namesB}{Names of the indices of total biomass.}

\item{namesP}{Names of the indices of proportion of recruits in biomass.}

\item{niter}{Number of iterations.}

\item{logscale}{Logical, if TRUE the parameters are in the scale used by the bbm model,
otherwise, all parameters are in the linear scale. This is used for example,
in the vcov matrix retuned from \code{bbm} function.}
}
\value{
An \code{FLPar} with the appropriate format for the \code{bbm} function.
}
\description{
This function generates an \code{FLPar} object with the input parametes required by the \code{bbm} function,
given information on the years, the indices names and the number of iterations.
}
\examples{

# Load data
data(ane)

years.ane <- dimnames(catch.ane)$year
niter.ane <- dim(catch.ane)[6]
namesB.ane <- names(indicesB.ane)
namesP.ane <- names(indicesP.ane)

# Generate population estimates, given some estimated parameters
pars <- bbmFLPar( years=years.ane, namesB=namesB.ane, namesP=namesP.ane, niter=niter.ane)
class(pars)
pars
 
}
\seealso{
\link{FLPar}, \link{bbm}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{bbmFLPar}
