% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBMpar.R
\docType{class}
\name{BBMpar}
\alias{BBMpar}
\alias{BBMpar-methods}
\alias{BBMpar-class}
\title{BBMpar class with information on parameters required by the BBM function.}
\arguments{
\item{object}{An \code{FLQuant} with catch information (for recruits and adults) or an \code{FLStock}.}

\item{indices}{Abundance indices in biomass for same age classes (class: FLQuants).}

\item{control}{List with other extra variables:
\itemize{
\item{logq   }{: Logarithm of the catchability parameter for each survey.}
\item{logpsi }{: Logarithm of the precisions of the observation equations of the total biomass index for each survey.}
\item{xi     }{: Parameter related to the variance of the observation equations of recruits biomass proportion for each survey.}
\item{logB0  }{: Logarithm of the total biomass at the begining of the first year.}
\item{logR   }{: Logarithm of annual recruitments at the beggining of the year.}
\item{mur    }{: Mean of the normal process for log(recruitment).}
\item{logpsir}{: Logarithm of the precision of the normal process for log(recruitment.}
}}
}
\value{
An object of class BBMpar.
}
\description{
BBMpar class with information on parameters required by the BBM function.
}
\section{Validity}{

\describe{
\item{XXX}{XXX}
}
You can inspect the class validity function by using
\code{getValidity(getClassDef('BBMpar'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If \code{nyear} or \code{nidex} are provided, this is used
for sizing the logR slot and the slots related to the indices (logq, logpsi and xi), respectively.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
\item{as.list}{: Converts the \code{BBMpar} object into a list.}
}
}

\examples{
# Load required libraries
library(bbm)

# Load data
data(ane)

# Generate an object of BBMpar class (different alternatives)
new("BBMpar")                # empty object
slotNames(BBMpar())          # slots

BBMpar( nyear=20, nindex=3)  # setting dimensions
BBMpar( logq=0)              # providing values
# BUG BBMpar( logq=0, xi=rep(5,3)) # but with some restrictions
# BUG BBMpar( logB0=rep(10,3))
# BUG BBMpar( logS=10)

# Available methods for the class
# - as.list()

l1 <- as.list(new("BBMpar"))
class(l1)

l2 <- base:::as.list(new("BBMpar"))
class(l2)

# Generate intitial values (output class is BBMpar)
inits <- initsBBM(catch.ane, indices=indices.ane, g=control.ane@g)
class(inits)

# Run assessment (inits must be of class BBMpar)
run <- BBM(catch.ane, indices=indices.ane, control=control.ane, inits=inits.ane)
run

}
\seealso{
\link{BBM}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{BBMpar}
