% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndices_fun.R
\name{calcIndices}
\alias{calcIndices}
\title{Function to estimate fitted indices}
\usage{
calcIndices(g, findicesB, findicesP, catch, inits)
}
\arguments{
\item{g}{A \code{vector} with information on instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult).}

\item{findicesB}{A \code{vector} with fraction of the year corresponding to each of the indicesB.}

\item{findicesP}{A \code{vector} with fraction of the year corresponding to each of the indicesP.}

\item{catch}{An \code{FLQuant} with catch information (for recruits and adults).}

\item{inits}{An \code{FLPar} with parameter values for the parameters required by the \code{bbm} function.}
}
\value{
A \code{list} with information on estimated indices (\code{FLQuants}).
The list has 2 elements: \code{indicesB} (for indices in biomass) and \code{indicesP} (for proportions of recruits in biomass).
}
\description{
This function estimates fitted indices, given information on growth, periods duration, catches and estimated parameters.
}
\section{Methods}{

Methods exist for various calculations based on the output class (\code{FLPar}). For details: \code{?FLPar}.
}

\examples{

# Load data
data(ane)

args(calcIndices)

indices <- calcIndices( g=control.ane@g, 
                        findicesB=unlist(lapply( indicesB.ane, function(x) mean(range(x)[c('startf','endf')]))),
                        findicesP=unlist(lapply( indicesP.ane, function(x) mean(range(x)[c('startf','endf')]))),
                        catch=catch.ane, inits=inits.ane )

class(indices)
slotNames(indices)

}
\seealso{
\link{FLPar}, \link{bbmFLPar}, \link{bbmControl}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{calcIndices}
