% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPop.R
\name{calcPop}
\alias{calcPop}
\title{Function to estimate abundances in mass at age by period.}
\usage{
calcPop(g, f, catch, inits)
}
\arguments{
\item{g}{A \code{vector} with information on instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult).}

\item{f}{A \code{vector} with fraction of the year corresponding to each of the periods (determined by the period of the year of the different indices).}

\item{catch}{An \code{FLQuant} with catch information (for recruits and adults).}

\item{inits}{An \code{BBMpar} with initial values for the parameters required by the BBM function.}
}
\value{
A \code{list(stock, ok)} with information on estimated stock (\code{FLQuant})
and an indicator on wheter estimated parameters are valid (i.e. positive, \code{ok==TRUE}) or not (\code{ok==FALSE}).
}
\description{
This function estimates abundances in mass at age (for recruits and adults) by period, given information on growth, periods duration, catches
and some additional values.
}
\examples{

# Load required libraries
library(bbm)

# Load data
data(ane)

# Generate population estimates, given some estimated parameters
bioAge <- calcPop(g=control.ane@g, f=periods(indices.ane)$f, catch=catch.ane, inits=inits.ane)
class(bioAge)

# check if valid ouput (i.e. positive biomass values)
bioAge$ok

# estimates
bioAge$stock
 

}
\seealso{
\link{BBMpar}, \link{BBM.control}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{calcPop}
