% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsBBM.R
\name{initsBBM}
\alias{initsBBM}
\alias{initsBBM-methods}
\title{initsBBM class for generating intial values.}
\usage{
initsBBM(object, indices, g)
}
\arguments{
\item{object}{An \code{FLQuant} with catch information (for recruits and adults) or an \code{FLStock}.}

\item{indices}{Abundance indices in biomass for same age classes (class: FLQuants).}

\item{g}{A \code{vector} with information on instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult).}
}
\value{
An object of class BBMpar.
}
\description{
Function to generate initial parameters required for the BBM function, given information on catches, survey indices
and instantaneous rate of biomass decrease, g.
}
\section{Methods}{

Methods exist for various calculations based on the output class (\code{BBMpar}). For details: \code{?BBMpar}.
}

\examples{

# Load required libraries
library(bbm)
library(ggplotFL)

# Load data
data(ane)

# Generate intitial values
inits <- initsBBM(object=catch.ane, indices=indices.ane, g=control.ane@g)
class(inits)

# Run assessment
run <- BBM(object=catch.ane, indices=indices.ane, control=control.ane, inits=inits)

# Plot assessed population
\dontrun{
  plot(stock(run))
}

}
\seealso{
\link{BBM}, \link{BBMpar}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{BBMpar}
