% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBM.R
\name{multiple_GSB}
\alias{multiple_GSB}
\title{BBM: A novel Beta-binomial-distribution-based Biclustering Algorithm for Mining m6A Co-methylation Patterns}
\usage{
multiple_GSB(a, b, iteration, burn_in, bic_row_labels, number_bicluster)
}
\arguments{
\item{a}{Ip matrix}

\item{b}{input matrix}

\item{iteration}{Number of initial iterations}

\item{burn_in}{Number of burnings}

\item{bic_row_labels}{Labels for sites in the patterns found, all with initial values of 0}

\item{number_bicluster}{Number of preset patterns}
}
\value{
the patterns hidden in the data
}
\description{
BBM: A novel Beta-binomial-distribution-based Biclustering Algorithm for Mining m6A Co-methylation Patterns
}
\examples{
rm(list = ls(all = TRUE))
library(BBM)
data("ip")
data("input")
number_bicluster <- 6
bic_row_labels <- c(rep(0,dim(ip)[1]))
patterns_found_by_BBM <- multiple_GSB(a = ip,b = input,iteration = 1000,burn_in = 500,bic_row_labels = bic_row_labels,number_bicluster = number_bicluster)









}
