% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods_fun.R
\name{periods}
\alias{periods}
\title{Function to estimate the seasons}
\usage{
periods(findicesB = NULL, findicesP = NULL)
}
\arguments{
\item{findicesB}{A named \code{vector} with fraction of the year when each of the total biomass indices are observed.}

\item{findicesP}{A named \code{vector} with fraction of the year when each of the proportion of recruits indices are observed.}
}
\value{
A \code{list} with four numeric elements: \code{nper}, with the number of seasons;
\code{f} with the fraction of the year corresponding to each of the seasons;
\code{perindicesB} with the season in which the index in total biomass is observed,
this object is a named vector with one element per index; and
\code{perindicesP} with the season in which the index of proportion of recruits is observed,
this object is a named vector with one element per index.
}
\description{
Given the fraction of the year when each of the indices are observed, this function estimates the number of seasons,
the fraction of the year of each season and the season when each of the indices are observed.
}
\examples{

# Load data
data(ane)

# Generate population estimates, given some estimated parameters
per <- periods( findicesB=unlist(lapply( indicesB.ane, function(x) mean(range(x)[c('startf','endf')]))), 
                findicesP=unlist(lapply( indicesP.ane, function(x) mean(range(x)[c('startf','endf')]))))
class(per)

per$nper
per$f
per$perindicesB
per$perindicesB

}
\seealso{
\link{}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{calcPop}
