% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmFit_class.R, R/bbmFitresiduals_class.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,bbmFit,missing-method}
\alias{plot,bbmFitresiduals,missing-method}
\title{Plot of Pearson residuals of survey indices}
\usage{
\S4method{plot}{bbmFit,missing}(x, y, ...)

\S4method{plot}{bbmFitresiduals,missing}(x, y = missing,
  auxline = "smooth", ...)
}
\arguments{
\item{x}{A \code{bbmFit} object with the fitted values and observed values.}

\item{y}{Missing.}

\item{...}{Additional argument list.}

\item{auxline}{A string defining the type of line to be added, by default uses 'smooth',
a common alternative is to use 'r', a regression, or leave it empty ''.}

\item{x}{An \code{bbmFitresiduals} object with the Pearson residuals.}

\item{y}{Ignored.}

\item{...}{Additional argument list.}
}
\value{
If \code{class(x)=='bbmFit'}, a \code{plot} with estimated abundances and one extra \code{plot} for each survey with fitted and observed indices.

If \code{class(x)=='bbmFitresiduals'}, a \code{plot} with stardardized residuals.
}
\description{
Method to plot bbm fitting, if applied to a \code{bbmFit} object. The output are several plots.
Firstly, the estimated abundance and next the fitted versus observed indices (with one plot for each survey).
Note that in plots related to indices the yaxis doesn't has a scale.
The visual is about the difference between the two lines, not about the value of each line, which in any case would be very difficult to assess visually.

Method to produce scatterplots of Pearson residuals of survey indices, if applied to a \code{bbmFitresiduals} object.
}
\examples{

# Load data
data(ane)

# Run the assessment
obj <- bbm( catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)

# Plot the output
plot(obj)


# Residuals
res <- residuals(obj)
plot(res)

}
