% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmFitresiduals_class.R
\docType{methods}
\name{qqplot}
\alias{qqplot}
\alias{qqmath}
\alias{qqmath,bbmFitresiduals,missing-method}
\title{Normal Q-Q Plot of surveys' indices}
\usage{
\S4method{qqmath}{bbmFitresiduals,missing}(x, data = missing, ...)
}
\arguments{
\item{x}{An \code{bbmFitresiduals} object with the Pearson residuals.}

\item{data}{Ignored.}

\item{...}{Additional argument list that might never be used.}
}
\value{
A qqplot with stardardized log residuals against a theoretical distribution.
}
\description{
Method to produce qqplots of Pearson residuals against theoretical distribution of surveys' indices.
}
\examples{

# Load data
data(ane)

# Run the assessment
obj <- bbm( catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)

# Residuals
res <- residuals(obj)
qqmath(res)

}
