% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBBM.R
\name{simBBM}
\alias{simBBM}
\alias{simBBM-methods}
\title{Function to generate indices.}
\usage{
simBBM(catch, g, inits, findex = NULL)
}
\arguments{
\item{catch}{An \code{FLQuant} with catch information (for recruits and adults).}

\item{g}{A \code{vector} with information on instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult).}

\item{inits}{An \code{BBMpar} with initial values.}

\item{f}{A \code{vector} with fraction of the year corresponding to each of the periods (determined by the period of the year of the different indices).}
}
\value{
An object of class FLInidices.
}
\description{
This function generates indices from a population (with only 2 age classes, recruits and adults), given information on catches at age,   and survey indices.
}
\examples{

# Load required libraries
library(bbm)
library(ggplotFL)

# Load data
data(ane)

# Generate intitial values
indices <- simBBM( catch=catch.ane, g=control.ane@g, inits=inits.ane, 
                   findex=unlist(lapply( indices.ane, function(x) (x@range[['startf']]+x@range[['endf']])/2)))
class(indices)

# Run assessment
run  <- BBM(catch.ane, indices=indices.ane, control=control.ane, inits=inits.ane)
run2 <- BBM(catch.ane, indices=indices, control=control.ane, inits=inits.ane)

# Plot assessed population
\dontrun{
  plot(FLQuants(bc=quantSums(run@stock.bio)[,,,1,], alt=quantSums(run2@stock.bio)[,,,1,]))
}

}
\seealso{
\link{BBM}, \link{BBMpar}
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{BBMpar}
