% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBMass.R
\name{BBM}
\alias{BBM}
\alias{BBM-methods}
\title{Function to simulate a two-stage biomass-based model.}
\usage{
BBM(object, indices, control, inits)
}
\arguments{
\item{object}{An \code{FLQuant} with catch information or an \code{FLStock}. These objects must only have two age classes (recruits and adults)
and the number of seasons should be 1 or the number of seasons determined by the period of the year of the different indices.}

\item{indices}{Abundance indices in biomass for same age classes (\code{FLIndices}).}

\item{control}{An \code{BBM.control} with control arguments.}

\item{inits}{An \code{BBMpar} with initial values.}
}
\value{
An object of class BBMfit.
}
\description{
BBM function simulates a two-stage biomass model. Specifically a maximum likelihood version of a generalisation of the model in Ibaibarriaga et al. (2008).
In Ibaibarriaga et al. (2008) only two indices ocurring at the same moment of the year were considered.
However, in this case is generalised to n indices that can occur in different times of the year.
}
\examples{

# Load required libraries
library(bbm)
library(ggplotFL)

# Load data
data(ane)

# Run assessment
run <- BBM(object=catch.ane, indices=indices.ane, control=control.ane, inits=inits.ane)

is(run)
names(run)

# Plot assesses population
\dontrun{
  plot(run@stock.bio)
}

}
\seealso{
\link{BBMfit}.
}
\author{
Leire Ibaibarriaga & Sonia Sanchez.
}
\keyword{BBM}
